allow displaying summary in post list
introduce a new `showSummary` boolean configuration variable that,
when `true`, causes the post summary to be displayed along with the
title in the list of posts.

the `shoreMore` boolean controls whether the summary is followed by a
"read more" prompt, and the `moreMarker` contains the actual text of
the link.